+++ lib/src/renderer/get_pixels/main.dart
@@ -1,9 +1,0 @@
-import 'dart:typed_data';
-
-import 'io.dart';
-
-Future<Uint8List> getPixels({
-  required String path,
-  bool removeTempFile = true,
-}) =>
-    getPlatformPixels(path: path, removeTempFile: removeTempFile);